backend/requirements.txt
Django>=3.0,<4.0
psycopg2>=2.8.6,<2.9
djangorestframework
django-cors-headers
tweepy
django-elasticsearch-dsl==7.2.0
django-elasticsearch-dsl-drf==0.22.2
requests
python-dateutil